Related work items: #5

+++ Services/Impls/AnuncioService.cs
@@ -142,9 +142,13 @@ namespace Services.Impls
         {
             if (await _temporizadorService.GroupHasTemporizadoresEnable(GrupoId))
             {
-                throw new Exception("Grupo tiene Temporizadores activados");
+                throw new CannotDeleteException("El Grupo tiene Temporizadores habilitados");
             }
-            IEnumerable<Anuncio> anuncios = await repositoryAnuncio.FindAllAsync(p => p.GroupId == GrupoId && p.Procesando == 0);
+            IEnumerable<Anuncio> anuncios = await repositoryAnuncio.FindAllAsync(p => p.GroupId == GrupoId);
+            if(anuncios.Any(a => a.Procesando != 0))
+            {
+                throw new CannotDeleteException("Existen anuncios siendo procesados en el sistema");
+            }
             repositoryAnuncio.RemoveRange(anuncios);
 
             await repositoryAnuncio.SaveChangesAsync();
@@ -188,7 +192,7 @@ namespace Services.Impls
         private IQueryable<Anuncio> QueryBaseAnunciosToUpdate(string GroupId)
         {
             return repositoryAnuncio.QueryAll().Where(a => a.GroupId == GroupId
-                                                        && a.Enable
+                                                        && a.Enable.HasValue && a.Enable.Value
                                                         && a.Procesando == 0
                                                         && !a.Eliminado)
                                                .OrderBy(a => a.Orden)
@@ -200,11 +204,11 @@ namespace Services.Impls
             Anuncio anuncio = await repositoryAnuncio.FindAsync(p => p.Id == Id);
             if(anuncio.Procesando != 0)
             {
-                throw new Exception("El anuncio está siendo procesado por el sistema");
+                throw new CannotDeleteException("El anuncio está siendo procesado por el sistema");
             }
             if (await _temporizadorService.GroupHasTemporizadoresEnable(anuncio.GroupId))
             {
-                throw new Exception("Grupo tiene Temporizadores activados");
+                throw new CannotDeleteException("El Grupo tiene Temporizadores activados");
             }
             repositoryAnuncio.Remove(anuncio);
             await repositoryAnuncio.SaveChangesAsync();
